refactoring and test

+++ atlasdb-client/src/main/java/com/palantir/atlasdb/stream/StreamStorePersistenceConfiguration.java
@@ -30,7 +30,7 @@ public interface StreamStorePersistenceConfiguration {
             .build();
 
     @Value.Default
-    default int numBlocksToWriteBeforePause() {
+    default long numBlocksToWriteBeforePause() {
         return 1;
     }
